Create a Storage interface to handle uploads
This creates a Storage interface (actually it is an abstract class due
to interfaces not existing in runtime) to handle file uploads.

For the time being, LocalStorage is its only implementation but having
the interface ready allows for rapid development of other
implementations.

e.g. in production we might want to use Amazon S3 for storage.