Create MatchLogsModule
This tries to aggregate MatchLog related dependencies in one place and
declutter the main AppModule.

Also, it is now possible to upload files and have it persisted to the
database. This was achieved by wiring the MatchLogModule to the
MatchLogsController.

Some files related to the MatchLogsController were renamed in an attempt
to stick to the naming conventions.