Rename some concepts to improve project readability
MatchLog was renamed to GameLog because it is actually related to an
entire game log file that was uploaded to the system. And now we have
temporary match log files while processing and that could lead to
confusion between games and matches.

Some other files that were named after MatchLog were also renamed
accordingly.

MatchLogStreamer and MatchLogConsumer were renamed to GameLogParser and
MatchLogParser. Again to improve awareness about games vs matches, but
also because Streamer and Consumer were not accurate names for what they
do. Parser is better at describing them.