Refactor AnalysisService to use SQL for its calculations
The calculations were being made with js, which was not ideal here since
we might be dealing with large datasets.

This should minimize the roundtrips to the database and also the amount
of data being exchanged in the network.

Also, the database engine should be far more efficient for those
calculations.